refactor: Tidy `features:resume:internal` module into packages
