+++ features/resume/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/root/ProveYourIdentityRootUiState.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.resume.internal
+package uk.gov.onelogin.criorchestrator.features.resume.internal.root
 
 internal data class ProveYourIdentityRootUiState(
     val shouldDisplay: Boolean,
