+++ features/resume/internal/src/test/snapshots/images/uk.gov.onelogin.criorchestrator.features.resume.internal_ScreenshotTest_screenshot[1.uk.gov.onelogin.criorchestrator.features.resume.internal.card.ProveYourIdentityUiCardKt_ProveYourIdentityUiCardPreview]_dark_night.png
